backend/requirements.txt
fastapi==0.70.1
uvicorn[standard]==0.15.0
python-dotenv==0.19.2
groq==0.11.0
httpx==0.23.0
python-multipart==0.0.5